requirements.txt
langchain
python-dotenv
langchain_community
pypdf
faiss-cpu
langchain_google_genai
pydantic
pymongo